refactor: ExecutionResult클래스 View패키지로 이동
InputOutputController에서 제어하는 ExecutionResult 클래스는 View패키지에 있는게 더 적합할거 같아 이동시켰다.
